+++ lab_agent/repositories/__init__.py
@@ -1,13 +1,17 @@
 """Repository interfaces for LabAgent persistence."""
 
 from .conversations import ConversationRepository
+from .email_login_challenges import EmailLoginChallengeRepository
 from .errors import OwnedResourceNotFoundError
 from .messages import MessageRepository
 from .users import UserRepository
+from .user_sessions import UserSessionRepository
 
 __all__ = [
     "ConversationRepository",
+    "EmailLoginChallengeRepository",
     "MessageRepository",
     "OwnedResourceNotFoundError",
     "UserRepository",
+    "UserSessionRepository",
 ]